Add support for IndieAuth Metadata endpoint
As part of a recent update to the IndieAuth spec, we've introduced the
`indieauth-metadata` endpoint, which allows us to push more
configuration for our service into a specific endpoint for
configuration, similar to RFC8414.

We should support this discovery for those who support it, falling back
to the legacy HTML/Link-only means when not supported.

We can use `FilterByRel` to catch the presence of the
`indieauth-metadata` and use that.

Closes #1.